refactor task_manager_server a little & implement run_task unit test
